simplified JpaRolloutManagement, moved rollout handling logic to JpaRolloutHandler
